requirements.txt
# Core dependencies
streamlit>=1.31.0
anthropic>=0.40.0
requests>=2.31.0

# PDF generation
fpdf2>=2.7.0
markdown>=3.5.0

# Additional utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.2
beautifulsoup4>=4.12.0
lxml>=4.9.0
pytrends>=4.9.0
howlongtobeatpy>=1.0.0

# Multi-model ensemble (Phase 3) - optional
openai>=1.0.0
google-generativeai>=0.3.0
